Add files via upload
I created two graphs which are forming a useful start to this portfolio and also shows the many interesting comparisons and analysis that can be made in this portfolio. First, i made a comparison of characteristic track features of two famous soundtrack composers: John Williams and Hans Zimmer. For both of them I used the ''This is'' playlist from Spotify. I made a visualization which shows if a song is major or minor, and showed the instrumentalness and valence levels for their tracks. The size of the circles show the energy level of the tracks. What you can clearly see already is that from both composers, most songs show high instrumentalness and low valence and this is most evident for Hans Zimmer. And he also has some tracks which score low in instrumentalness and low on valence, which I called ''sombre vocalization''. But he also has some outliers which are high on valence, on a level that the outliers of John Williams don't reach. This is what I called ''happy instrumentation''.  It is also clear that the majority is written in minor and that a large level of songs has an average energy level.

Secondly I made a scatterplot from the playlist ''Iconic Soundtracks'', which also consists soundtracks of various other composers and here I compared instrumentalness and energy. Here it is clearly shown again that most tracks score high on instrumentalness and have quite low energy levels. Again, certain outliers on both levels still exist. 
What I think is interesting from these analyses is that it shows that even two composers show quite some variety in the characteristics of their tracks, but that roughly speaking there are also similarities. I am interested to further investigate comparisons from other big composers and from main famous soundtracks in general, to find out more interesting differences and similarities. And I am also interested to further analyse specific outliers, to find out what is so different about them.